feat: add catalog games template. refactor: templates
