+++ src/components/templates/shared/MainLayout.tsx
@@ -7,7 +7,7 @@ interface MainLayoutProps {
 
 export default function MainLayout({ children }: MainLayoutProps) {
   return (
-    <div className="flex flex-col min-h-screen">
+    <div className="flex flex-col min-h-screen min-w-56">
       <Header />
       <main className="flex-grow">{children}</main>
       <Footer />
